Rename file & move function to util
