+++ backend/tests/unit_tests/shopping_cart/test_route.py
@@ -8,11 +8,8 @@ from sqlalchemy import func
 
 from src.database import db
 from src.models import Item, ShoppingCart, Tag, TagOfItem
-from src.shopping_cart.route import fetch_user_id_from_jwt_token
 
 if TYPE_CHECKING:
-    from http.cookiejar import Cookie, CookieJar
-
     from flask import Flask
     from flask.testing import FlaskClient
     from werkzeug.test import TestResponse
@@ -79,20 +76,6 @@ def setup_item(app: Flask) -> None:
             ],
         )
         db.session.commit()
-
-
-def test_user_id_fetch_function_with_logged_in_client_should_return_user_id(
-    app: Flask, logged_in_client: FlaskClient
-) -> None:
-    with app.app_context():
-        cookies: tuple[Cookie, ...] = _get_cookies(logged_in_client.cookie_jar)
-        (jwt_cookie,) = tuple(filter(lambda x: x.name == "jwt", cookies))
-        assert jwt_cookie.value is not None
-        jwt_token = jwt_cookie.value
-
-        user_id: int = fetch_user_id_from_jwt_token(jwt_token)
-
-        assert user_id == 1  # The user with test@mail.com UID is 1.
 
 
 class TestGetShoppingCart:
@@ -334,9 +317,3 @@ class TestPutShoppingCartItem:
         )
 
         assert response.status_code == HTTPStatus.FORBIDDEN
-
-
-def _get_cookies(cookie_jar: CookieJar | None) -> tuple[Cookie, ...]:
-    if cookie_jar is None:
-        return tuple()
-    return tuple(cookie for cookie in cookie_jar)
